Average from 3 runs. Tested on node 9.7.1


benchmark | avg old [ms] | avg new [ms] | % diff [100 - old/new * 100 %]
-- | -- | -- | --
simple url | 2274.382 | 2031.633 | -11.94846707
complex url | 11664.32533 | 10741.85067 | -8.587669809
short query | 8876.287667 | 7680.794333 | -15.5647096
long query | 27470.93667 | 20070.11033 | -36.87486621