@aldor007 2018-10-15
@albert-lacki @kuebk @mgren @krzys1u @psulowicz 
**ping**
![](https://media.giphy.com/media/C2ZLKN9SSOFmU/giphy.gif)